Address Mahesh's comments
This patch combines tile and winograd input and
output patterns to 1 pass.
