+++ compiler/src/iree/compiler/Codegen/LLVMCPU/Passes.cpp
@@ -142,9 +142,7 @@ static void addTileAndDistributePasses(
   nestedModulePM.addPass(createCanonicalizerPass());
   nestedModulePM.addPass(createCSEPass());
   nestedModulePM.addNestedPass<func::FuncOp>(
-      IREE::LinalgExt::createTileAndDecomposeWinogradInputTransformPass());
-  nestedModulePM.addNestedPass<func::FuncOp>(
-      IREE::LinalgExt::createTileAndDecomposeWinogradOutputTransformPass());
+      IREE::LinalgExt::createTileAndDecomposeWinogradTransformPass());
 }
 
 //===---------------------------------------------------------------------===//
