+++ llvm-external-projects/iree-dialects/lib/Dialect/LinalgExt/Passes/CMakeLists.txt
@@ -5,8 +5,7 @@ add_mlir_library(IREELinalgExtPasses
   PadContractionToBlockSize.cpp
   Passes.cpp
   SplitReduction.cpp
-  TileAndDecomposeWinogradInputPass.cpp
-  TileAndDecomposeWinogradOutputPass.cpp
+  TileAndDecomposeWinogradPass.cpp
   Tiling.cpp
 
   DEPENDS
